move relevant tests to their own module
